Updated files

+++ digi_analysis/functions.cpp
@@ -1,8 +1,11 @@
+// Copyright (c) 2025
+//
 // This file contains hand‑reconstructed C functions that approximate
 // the behaviour of some of the simple routines at the very beginning of
 // digi.exe.  These functions were derived by inspecting the disassembly
-// at addresses 0x401000, 0x401020, 0x401040 and 0x401050.  Each function
-// operates on an integer input and looks up a 16‑bit value in a table.
+// at addresses 0x401000, 0x401020, 0x401040 and 0x401050.  Each
+// function operates on an integer input and looks up a 16‑bit value
+// in a table extracted from the game binary.
 
 #include <cstdint>
 #include "digi_table.h"
@@ -12,14 +15,14 @@
 // returns the 16‑bit value.  The original used movsx to sign
 // extend the 16‑bit value to 32 bits; here we return a 16‑bit int.
 extern "C" int16_t func_401000(int32_t value) {
-    int index = value & 0xFFF;
+    int index = static_cast<int>(value) & 0xFFF;
     return g_table[index];
 }
 
 // Corresponds to the function at address 0x401020.  It adds 0x400 to
 // the argument, masks to 12 bits and indexes the same table.
 extern "C" int16_t func_401020(int32_t value) {
-    int index = (value + 0x400) & 0xFFF;
+    int index = static_cast<int>(value + 0x400) & 0xFFF;
     return g_table[index];
 }
